CODE RUB: Add NumericeEntry.xaml.Properties,cs Closes #20
